add jupiter file

+++ backend/app/schemas/predict.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 
 class PlacementRequest(BaseModel):
-    cgpa: int = Field(..., ge=0, le=10, description="CGPA must be between 1 and 10")
+    cgpa: float = Field(..., ge=0, le=10, description="CGPA must be between 1 and 10")
     internship: int = Field(..., ge=0, le=10, description="Internship must be between 0 and 10")
 
 class PlacementResponse(BaseModel):
